+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## unrelease
+* Fix bug where SDK expects `granted_payment_instrument_revoked` instead of `granted_payment_method_revoked`
+
 ## 2.88.0
 * Add `travel_flight` support to industry-specific data